add point component
