+++ frontend/src/constants.ts
@@ -1,0 +1,16 @@
+// real latitude coordinate range
+export const mapMinX = -155.905;
+export const mapMaxX = -155.58;
+
+// real longitude coordinate range
+export const mapMinY = 20.01;
+export const mapMaxY = 20.27;
+
+// map and table aspect ratios
+export const aspect5_4 = "aspect-[calc(5/4)]";
+export const aspect16_9 = "aspect-[calc(16/9)]";
+
+// calculations for web
+export const mapRealWidthInches = 29.5;
+export const tableRealWidthInches = 42.666;
+export const mapWidthPercent = `w-[calc(${mapRealWidthInches}/${tableRealWidthInches}*100%)]`;
