+++ frontend/src/types.ts
@@ -1,0 +1,4 @@
+// [long, lat] follows the GeoJSON convention
+export type Longitude = number;
+export type Latitude = number;
+export type Coordinates = [Longitude, Latitude];
