--- map-app/src/Controller.tsx
@@ -1,60 +1,0 @@
-import { Button } from "@mui/material";
-import { controlMap, zoomIn, zoomOut } from "./api";
-
-import { useMutation } from "@tanstack/react-query";
-import { CloseButton } from "./ui/CloseButton";
-import { useNavigate } from "react-router";
-
-export function Controller() {
-  const navigate = useNavigate();
-  const mutation = useMutation({
-    mutationFn: controlMap,
-    onSuccess: () => {},
-  });
-
-  const mutationZoomIn = useMutation({
-    mutationFn: zoomIn,
-  });
-  const mutationZoomOut = useMutation({
-    mutationFn: zoomOut,
-  });
-
-  return (
-    <div className="m-4">
-      <div className="w-full flex justify-between mb-4">
-        <div className="text-2xl mb-4">Controller</div>
-        <CloseButton
-          onClick={() => {
-            navigate("/");
-          }}
-        />
-      </div>
-
-      {/* <Button
-        onClick={() => {
-          mutation.mutate({ layers: ["Test!"] });
-        }}
-      >
-        Layers
-      </Button> */}
-      <div className="flex flex-col gap-4">
-        <Button
-          variant="contained"
-          onClick={() => {
-            mutationZoomIn.mutate();
-          }}
-        >
-          Zoom In
-        </Button>
-        <Button
-          variant="contained"
-          onClick={() => {
-            mutationZoomOut.mutate();
-          }}
-        >
-          Zoom Out
-        </Button>
-      </div>
-    </div>
-  );
-}
