--- map-app/src/vite-env.d.ts
@@ -1,2 +1,0 @@
-/// <reference types="vite/client" />
-/// <reference types="@arcgis/map-components/types/react" />
