--- map-app/vite.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
-import tailwindcss from "@tailwindcss/vite";
-
-// https://vite.dev/config/
-export default defineConfig({
-  server: {
-    port: 5173,
-  },
-  plugins: [react(), tailwindcss()],
-});
